requirements.txt
fastapi
uvicorn
torch
sentence-transformers
faiss-cpu
openai
python-dotenv
pymupdf
streamlit
pypdf
numpy
